+++ R/PrepInit.R
@@ -11,22 +11,40 @@
 #'
 #' @param verbose If \code{TRUE} then \code{message} progress information.
 #'
-#' @param gamma.alpha Shape parameter of gamma distribution from which
-#'   the Dirichlet process concentration parameters are drawn; in this
-#'   function the gamma distributions for all Dirichlet processes are the same.
+#' @param gamma.alpha Shape parameter of
+#'   the gamma distribution prior for the Dirichlet process concentration
+#'   parameters; in this
+#'   function the gamma distributions for all Dirichlet processes,
+#'   except possibly the top level process, are the same.
 #'
-#' @param gamma.beta Inverse scale parameter (rate parameter) of gamma distribution
-#'   from which the Dirichlet process concentration parameters are drawn; in this
-#'   function the gamma distributions for all Dirichlet processes are the same.
+#' @param gamma.beta Inverse scale parameter (rate parameter) of
+#'   the gamma distribution prior for the Dirichlet process concentration
+#'   parameters; in this
+#'   function the gamma distributions for all Dirichlet processes, except
+#'   possibly the top level process, are the same.
+#'
+#' @param gamma0.alpha See figure B.1 from Nicola Robert's thesis.
+#'   The shape parameter (\eqn{\alpha_0}) of the gamma
+#'   distribution priors for
+#'   the Dirichlet process concentration parameters  (\eqn{\gamma_0})
+#'   for \eqn{G_0}.
+#'
+#' @param gamma0.beta See figure B.1 from Nicola Robert's thesis.
+#'   Inverse scale parameter (rate parameter, \eqn{\beta_0}) of the gamma
+#'   distribution priors for
+#'   the Dirichlet process concentration parameters (\eqn{\gamma_0})
+#'   for \eqn{G_0}.
 #'
 #' @export
 
 PrepInit <- function(multi.types,
                      input.catalog,
-                     verbose,
+                     verbose      = TRUE,
                      K.guess,
-                     gamma.alpha=1,
-                     gamma.beta=1){
+                     gamma.alpha  = 1,
+                     gamma.beta   = 1,
+                     gamma0.alpha = gamma.alpha,
+                     gamma0.beta  = gamma.beta){
 
   if (mode(input.catalog) == "character") {
     if (verbose) message("Reading input catalog file ", input.catalog)
@@ -59,8 +77,8 @@ PrepInit <- function(multi.types,
             " rate (inverse scale) = ", gamma.beta)
   }
 
-  alphaa <- rep(gamma.alpha,dp.levels)
-  alphab <- rep(gamma.beta,dp.levels)
+  alphaa <- c(gamma0.alpha, rep(gamma.alpha, dp.levels - 1))
+  alphab <- c(gamma0.beta,  rep(gamma.beta,  dp.levels - 1))
 
   invisible(list(num.tumor.types = ppindex$num.tumor.types,
                  number.channels = number.channels,