+++ R/SetupAndActivate.R
@@ -15,7 +15,9 @@ SetupAndActivate <- function(input.catalog,
                              multi.types         = FALSE,
                              verbose             = TRUE,
                              gamma.alpha         = 1,
-                             gamma.beta          = 1)
+                             gamma.beta          = 1,
+                             gamma0.alpha        = gamma.alpha,
+                             gamma0.beta         = gamma.beta)
 { # 7 arguments
 
     prep_val <- PrepInit(multi.types = multi.types,
@@ -23,7 +25,9 @@ SetupAndActivate <- function(input.catalog,
                          verbose       = verbose,
                          K.guess       = K.guess,
                          gamma.alpha   = gamma.alpha,
-                         gamma.beta    = gamma.beta)
+                         gamma.beta    = gamma.beta,
+                         gamma0.alpha  = gamma0.alpha,
+                         gamma0.beta   = gamma0.beta)
 
     if (verbose) message("calling hdp_init ", Sys.time())
     hdpObject <- hdpx::hdp_init(ppindex = prep_val$ppindex,